convert workdays type from double to int

+++ src/main/scala/de/codecentric/ccdashboard/service/timesheet/data/access/ReportAggregator.scala
@@ -68,7 +68,7 @@ class ReportAggregator(reports: List[(Date, ReportEntry)], workSchedule: List[Us
     val valuesByKey = requiredHoursByKey.map({ case (id, requiredHours) =>
       val report = hoursReportByKey.getOrElse(id, ReportEntry())
       val utilizationValue = utilization(requiredHours, report.billableHours.getOrElse(0.0))
-      val workDays = requiredHours / 8
+      val workDays = (requiredHours / 8).toInt
       // Note: Other indicators may be inserted here
       id -> (report, utilizationValue, workDays)
     })
@@ -90,7 +90,7 @@ object ReportAggregator {
   private val yearFormatter = DateTimeFormatter.ofPattern("yyyy")
 }
 
-case class ReportAggregation(key: String, report: ReportEntry, utilization: Double, workDays:Double, numberOfConsultants: Int = 1)
+case class ReportAggregation(key: String, report: ReportEntry, utilization: Double, workDays: Int, numberOfConsultants: Int = 1)
 
 case class ReportAggregationResult(overallHoursRequired: Double,
                                    overallBillableHours: Double,
